+++ NEWS.md
@@ -1,5 +1,15 @@
 # inlmisc 0.4.4.9000
 
+- Add **alphahull** (released under a restrictive non-free software license)
+  and **maptools** to suggested packages.
+
+- Remove **httr** from suggested packages.
+
+- In `GetRegionOfInterest` function: add `alpha` and `width` arguments,
+  used to compute alpha-shape and expand the region of interest, respectively.
+
+# inlmisc 0.4.4
+
 - Add `BuildVignettes` function, used to build package vignettes.
 
 - In `SummariseBudget` function, improve memory management.